Getting close to stable combat.
Fixed a bug that caused the host not to execute TurnEnded event.
Fixed both players using their own energy resource bar.
Fixed parry and perfect parry not absorbing Skill StatusEffects.

Added/Changed skills.
Renamed Block to Dodge.
Renamed logging scripting def.
Made Passive energy gain be at end of turn instead of start.

+++ Assets/C# Scripts/DataTypes/MatchSettings.cs
@@ -4,6 +4,6 @@
 [System.Serializable]
 public struct MatchSettings
 {
-    public float PassiveEnergyGain;
+    public int PassiveEnergyGain;
     public PerfectParryPenaltyStats PerfectParryPenalty;
 }
